+++ core/Updates/1.5-b4.php
@@ -17,7 +17,7 @@ use Piwik\Updates;
  */
 class Updates_1_5_b4 extends Updates
 {
-    public static function getSql()
+    public function getMigrationQueries(Updater $updater)
     {
         return array(
             'ALTER TABLE `' . Common::prefixTable('site') . '`
@@ -25,8 +25,8 @@ class Updates_1_5_b4 extends Updates
         );
     }
 
-    public static function update()
+    public function doUpdate(Updater $updater)
     {
-        Updater::updateDatabase(__FILE__, self::getSql());
+        $updater->executeMigrationQueries(__FILE__, $this->getMigrationQueries($updater));
     }
 }