+++ core/Updates/2.10.0-b5.php
@@ -43,7 +43,7 @@ class Updates_2_10_0_b5 extends Updates
 {
     public static $archiveBlobTables;
 
-    public static function getSql()
+    public function getMigrationQueries(Updater $updater)
     {
         $sqls = array('# ATTENTION: This update script will execute some more SQL queries than that below as it is necessary to rebuilt some archives #' => false);
 
@@ -99,9 +99,9 @@ class Updates_2_10_0_b5 extends Updates
         return $sqls;
     }
 
-    public static function update()
+    public function doUpdate(Updater $updater)
     {
-        Updater::updateDatabase(__FILE__, self::getSql());
+        $updater->executeMigrationQueries(__FILE__, $this->getMigrationQueries($updater));
 
         // DeviceDetection upgrade in beta1 timed out on demo #6750
         $archiveBlobTables = self::getAllArchiveBlobTables();