+++ core/Updates/2.10.0-b8.php
@@ -10,10 +10,11 @@
 namespace Piwik\Updates;
 
 use Piwik\Updates;
+use Piwik\Updater;
 
 class Updates_2_10_0_b8 extends Updates
 {
-    public static function update()
+    public function doUpdate(Updater $updater)
     {
         $pluginManager = \Piwik\Plugin\Manager::getInstance();