+++ core/Updates/2.4.0-b1.php
@@ -9,10 +9,11 @@
 namespace Piwik\Updates;
 
 use Piwik\Updates;
+use Piwik\Updater;
 
 class Updates_2_4_0_b1 extends Updates
 {
-    public static function update()
+    public function doUpdate(Updater $updater)
     {
         try {
             \Piwik\Plugin\Manager::getInstance()->activatePlugin('Morpheus');