+++ core/Updates/2.5.0-b1.php
@@ -10,10 +10,11 @@ namespace Piwik\Updates;
 
 use Piwik\Config;
 use Piwik\Updates;
+use Piwik\Updater;
 
 class Updates_2_5_0_b1 extends Updates
 {
-    public static function update()
+    public function doUpdate(Updater $updater)
     {
         self::updateConfig();
     }